Update ConnectionConfig to simply rely on a URL instead of separate host, port, database name and driver
